+++ CHANGELOG.md
@@ -5,12 +5,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add compat option to generate narrowable union types (no parent is generated) `--compat-narrowable-unions`
+
 ## [6.0.3-beta.20230608.1] - 2023-06-08
 
 ### Added
 
--   Initial forked version
+- Initial forked version
 
 [Unreleased]: https://github.com/dolly22/autorest.typescript.compat/compare/v6.0.3-beta.20230608.1...HEAD
-
 [6.0.3-beta.20230608.1]: https://github.com/dolly22/autorest.typescript.compat/compare/63e07b0fdc6614be5c721f07d68d515a3ccab0ff...v6.0.3-beta.20230608.1